[APM] Relocate metrics routes (#40850)
* [APM] Relocate metrics routes

* fix tests
