+++ x-pack/legacy/plugins/apm/public/components/app/ServiceMetrics/MetricsChart.tsx
